Update to method for reading PGO source data

+++ prepare_data/prep_asc.py
@@ -5,26 +5,27 @@
 import os
 import pandas as pd
 import datetime
+import pickle
 import geopandas as gp
 
 from utils import pgo
 
 # Run for all ascii files in vaatplanten
 vaatplant_dir = r'd:\hotspot_working\b_vaatplanten\Soortenrijkdom'
-holder = []
 i = 0
-with open(os.path.join(vaatplant_dir, 'vaatplant_all4.csv'), 'w') as f:
+with open(os.path.join(vaatplant_dir, 'vaatplant_all_snl.csv'), 'w') as f:
     for file in os.listdir(vaatplant_dir):
-        if file.endswith('asc') and file.startswith('vaatplant') and os.path.isfile(os.path.join(vaatplant_dir, file)):
+        if file.endswith('asc') and file.startswith('vaatplant') and os.path.isfile(os.path.join(vaatplant_dir, file))\
+                and 'SNL' in file:
             print('{0} in progress'.format(file))
             foo = pgo.ascii_species_grid_to_pd(vaatplant_dir, file)
             print('\tDone with {0} rows at {1}'.format(foo.shape[0], datetime.datetime.now().strftime("%H:%M:%S")))
             # holder.append(foo)
             if i == 0:
-                f.write('# Deze tabel bevat alle informatie uit alle *.asc bestanden '
+                f.write('# Deze tabel bevat alle informatie uit alle *SNL*.asc bestanden '
                         'in d:\hotspot_working\b_vaatplanten\Soortenrijkdom\n')
                 f.write('# Deze bestanden zijn gebasseerd op de data levering van Kampichler 11 juni 2019\n')
-                f.write('# Hans Roelofsen, WEnR BB, 12 Jun 2019 12:27')
+                f.write('# Hans Roelofsen, WEnR BB, 04 Jul 2019\n')
                 f.write(foo.to_csv(sep=';', header=True, index=False))
             else:
                 f.write(foo.to_csv(sep=';', header=False, index=False))
